requirements.txt
ane-transformers==0.1.1
flagembedding==1.3.2
fqdn==1.5.1
google==3.0.0
google-cloud-texttospeech==2.21.0
google-generativeai==0.8.3
greenlet==3.1.1
grpcio-status==1.67.0
isoduration==20.11.0
jsonlines==4.0.0
jupyter==1.1.1
langchain-cohere==0.3.1
langdetect==1.0.9
llama-index==0.11.22
pinecone-client==5.0.1
pip-chill==1.0.3
pip-review==1.3.0
pipreqs==0.5.0
pyttsx3==2.98
ragas==0.2.4
ray==2.38.0
sounddevice==0.5.1
soundfile==0.12.1
streamlit==1.40.1
uri-template==1.3.0
watchdog==6.0.0
webcolors==24.8.0
